tests/wpt/tests/html/canvas/offscreen/canvas-context/2d.canvas.context.unrecognised.nullsuffix.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.canvas.context.unrecognised.nullsuffix</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.canvas.context.unrecognised.nullsuffix</h1>
<p class="desc">Context name "2d" plus a "\0" suffix is unrecognised</p>


<script>
var t = async_test("Context name \"2d\" plus a \"\\0\" suffix is unrecognised");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  var offscreenCanvas2 = new OffscreenCanvas(100, 50);
  assert_throws_js(TypeError, function() { offscreenCanvas2.getContext("2d\0"); });
  t.done();

});
</script>
